refactor: Pawn은 WhitePawn과 BlackPawn으로 분리하여 관리
