+++ README.md
@@ -54,4 +54,5 @@
 - [x] 기물(Piece) 객체가 상태로 방향 객체를 가지게 하기
 - [x] 기물에서 이동경로 생성 책임지기
 - [x] 폰의 공격 움직임 책임 수정
+- [x] 폰의 white와 black 각각으로 관리
 - [ ] 움직임에 대한 검증 로직을 더 세분화 하기
